Move some logic from TargetCard into model-like classes
The original idea was to persist all data from TargetCard, but turns out only two things are writable (which was obvious from the start, btw), and the end result is more or less the same, just in a different way lol But now persistence is only done from models, at least. Would probably make it a bit more organized to move to a cloud storage.